Remove unused files Move error types to common file
